refactor: created custom hook for progress bar movement
