+++ app/(pages)/about-us/_components/Archive/components/Carousel/carouselSubComponents/CarouselProgress/useProgress.tsx
@@ -1,0 +1,34 @@
+'use client';
+import { useState, useEffect, useCallback } from 'react';
+import { EmblaCarouselType } from 'embla-carousel';
+
+export const useProgress = (emblaApi: any) => {
+  // used to move the progress bar as the user moves through the carousel
+  const [scrollProgress, setScrollProgress] = useState(0);
+
+  const onScroll = useCallback((emblaApi: EmblaCarouselType) => {
+    const progress = Math.max(0, Math.min(1, emblaApi.scrollProgress()));
+    setScrollProgress(progress);
+  }, []);
+
+  useEffect(() => {
+    if (!emblaApi) return;
+
+    onScroll(emblaApi);
+    emblaApi.on('reInit', onScroll);
+    emblaApi.on('scroll', onScroll);
+  }, [emblaApi, onScroll]);
+
+  const handleProgressBarClick = (event: any) => {
+    const progressBarWidth = event.target.offsetWidth;
+    const clickedProgressPixels = event.nativeEvent.offsetX;
+    const clickedProgressPercent = clickedProgressPixels / progressBarWidth;
+
+    if (!emblaApi) return;
+    const snapList = emblaApi.scrollSnapList();
+    const indexToMoveTo = Math.floor(snapList.length * clickedProgressPercent);
+    emblaApi.scrollTo(indexToMoveTo);
+  };
+
+  return { scrollProgress, handleProgressBarClick };
+};
